Went on a side mission, updated our docs styling and created a dropdown menu for api reference
