+++ config.yml
@@ -8,7 +8,7 @@ validator:
   staked: "100000000stake"
 client:
   openapi:
-    path: "docs/static/blockchain-api.yml"
+    path: "docs/static/blockchain.yaml"
   vuex:
     path: "web/wallet/src/store"
 faucet:
