+++ docs/src/components/HomepageFeatures.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import clsx from 'clsx';
-import styles from './HomepageFeatures.module.css';
 
 const FeatureList = [
   {
@@ -36,11 +35,11 @@ function Feature({ Svg, title, description }) {
   return (
     <div className={clsx('col col--4')}>
       <div className="text--center">
-        <Svg className={styles.featureSvg} alt={title} />
+        <Svg className="featureSvg" alt={title} />
       </div>
       <div className="text--center padding-horiz--md">
-        <h3>{title}</h3>
-        <p>{description}</p>
+        <h3 className='typographyRichTextHeading3'>{title}</h3>
+        <p className='typographyRichTextParagraphDefault'>{description}</p>
       </div>
     </div>
   );
@@ -48,7 +47,7 @@ function Feature({ Svg, title, description }) {
 
 export default function HomepageFeatures() {
   return (
-    <section className={styles.features}>
+    <section className="features">
       <div className="container">
         <div className="row">
           {FeatureList.map((props, idx) => (
