+++ docs/src/css/colors.css
@@ -1,0 +1,18 @@
+
+:root {
+    --colorPrimaryLight: #1792ff;    
+    --colorPrimaryDark: #046de8;    
+    --colorSecondary1: #9665fd;    
+    --colorSecondary2: #1d1a27;    
+    --colorSecondarySuccess: #14b69a;    
+    --colorSecondaryError: #ff2866;    
+    --colorNeutral800: #13111a;    
+    --colorNeutral700: #302c3f;    
+    --colorNeutral600: #5e5b71;    
+    --colorNeutral500: #88849c;    
+    --colorNeutral400: #b7b4c7;    
+    --colorNeutral300: #d9d7e6;    
+    --colorNeutral200: #f5f4fa;    
+    --colorNeutral100: #ffffff;    
+    --colorSecondary3: #00bfff;    
+}
