+++ internal/host/ipfs/protocol.go
@@ -18,13 +18,13 @@ import (
 // IPFSProtocol leverages the IPFSLite library to provide simple file operations.
 type IPFSProtocol struct {
 	ctx       context.Context
-	node      host.HostImpl
+	node      host.SonrHost
 	dataStore *MemoryStore
 	*ipfslite.Peer
 }
 
 // New creates a new IPFSProtocol instance with Host Implementation
-func New(ctx context.Context, host host.HostImpl) (*IPFSProtocol, error) {
+func New(ctx context.Context, host host.SonrHost) (*IPFSProtocol, error) {
 	// Create IPFS Peer
 	ds := NewMemoryStore()
 	ipfsLite, err := ipfslite.New(ctx, ds.Batching(), host.Host(), host.Routing(), nil)
